+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+#### master
+
+- SERVER: Created .QUEBEC
+
+- SERVER: Updated .CEO, .CLUB, .VG
+
+
 #### Release 3.5.3
 
 - CHANGED: Updated whois.pir.org to the new response format.