+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 - CHANGED: Updated ccwhois.ksregistry.net to the new response format.
 
+- CHANGED: Updated whois.gg to the new response format.
+
 
 #### Release 3.5.3